requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-dotenv==1.0.1
requests==2.32.3
certifi==2025.1.31
urllib3==2.3.0
pydantic==2.10.6
python-multipart==0.0.20
starlette==0.45.3
typing-extensions==4.12.2
jinja2==3.1.5
aiofiles==24.1.0
gTTS==2.5.4 